+++ infrastructure-layer/src/main/java/com/wanderfun/infrastructurelayer/service/places/PlaceCategoryServiceImpl.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class PlaceCategoryServiceImpl extends BaseServiceImpl<PlaceCategory, Integer> implements PlaceCategoryService {
+public class PlaceCategoryServiceImpl extends BaseServiceImpl<PlaceCategory, Long> implements PlaceCategoryService {
     private final PlaceCategoryRepository placeCategoryRepository;
 
     @Autowired
@@ -36,7 +36,7 @@ public class PlaceCategoryServiceImpl extends BaseServiceImpl<PlaceCategory, Int
 
     @Override
     @Transactional
-    public PlaceCategory findById(Integer id) {
+    public PlaceCategory findById(Long id) {
         return placeCategoryRepository.findById(id)
                 .orElseThrow(() -> new ObjectNotFoundException(String.format("%s not found", PlaceCategory.class.getName())));
     }
